refactor for understandablility
