+++ src/geometry/primitives/plane.cc
