+++ src/geometry/primitives/sphere.cc
