+++ src/geometry/primitives/triangle.cc
