+++ src/geometry/primitives/triangle.h
