+++ tests/cow_reflect_refract_test.cc
@@ -2,7 +2,6 @@
 #include <iostream>
 
 #include "array_ops.h"
-#include "bvh.h"
 #include "camera.h"
 #include "color.h"
 #include "image.h"
@@ -12,9 +11,10 @@
 #include "line.h"
 #include "linear_types.h"
 #include "material.h"
+#include "multitudes/bvh.h"
 #include "parser.h"
-#include "plane.h"
-#include "sphere.h"
+#include "primitives/plane.h"
+#include "primitives/sphere.h"
 #include "vector_ops.h"
 
 using namespace linalg;
